fix error on logout and handle 0 plant locations
