+++ src/features/user/Account/styles/TreeMapper.module.scss
@@ -250,7 +250,7 @@
     > .mapContainer {
       // width: 50%;
       > .map {    
-      width: 50%;
+      width: 46%;
       }
     }
   }
